use decorator
